refactor: move sobrecontent component to its own folder
