+++ src/screens/public/PartnerProfile/SobreContent/index.ts
@@ -1,0 +1,1 @@
+export { SobreContent } from './SobreContent';
